+++ modules/@apostrophecms/doc-type/index.js
@@ -193,6 +193,9 @@ module.exports = {
     if (!self.options.name) {
       self.options.name = self.__meta.name;
     }
+    if (self.options.replicate === undefined) {
+      self.options.replicate = self.options.localized && self.options.singleton?.auto;
+    }
     self.name = self.options.name;
     // Each doc-type has an array of fields which will be updated
     // if the document is moved to the archive. In most cases 'slug'