+++ modules/@apostrophecms/doc/index.js
@@ -1,6 +1,7 @@
 const _ = require('lodash');
 const cuid = require('cuid');
 const { SemanticAttributes } = require('@opentelemetry/semantic-conventions');
+const { klona } = require('klona');
 
 // This module is responsible for managing all of the documents (apostrophe "docs")
 // in the `aposDocs` mongodb collection.
@@ -296,6 +297,144 @@ module.exports = {
   },
   methods(self) {
     return {
+      // `pairs` is an array of arrays, each containing an old _id
+      // and a new _id that should replace it.
+      //
+      // `aposDocId` is implicitly updated, `path` is updated if a page,
+      // and all references found in relationships are updated via reverse
+      // relationship id lookups, after which attachment references are updated.
+      // This is a slow operation, which is why this method should be called only
+      // by migrations and tasks that remedy an unexpected situation. _id is
+      // meant to be an immutable property, this method is a workaround
+      // for situations like a renamed locale or a replication bug fix.
+      //
+      // If `keep` is set to `'old'` the old document's content wins
+      // in the event of a conflict. If `keep` is set to `'new'` the
+      // new document's content wins in the event of a conflict.
+      // If `keep` is not set, a `conflict` error is thrown in the
+      // event of a conflict.
+
+      async changeDocIds(pairs, { keep }) {
+        let kept = 0;
+        // Get page paths up front so we can avoid multiple queries when working on path changes
+        const pages = await self.apos.doc.db.find({
+          path: { $exists: 1 },
+          slug: /^\//
+        }).project({
+          path: 1,
+          aposLastTargetId: 1
+        }).toArray();
+        for (const pair of pairs) {
+          const [ from, to ] = pair;
+          const existing = await self.apos.doc.db.findOne({ _id: from });
+          if (!existing) {
+            throw self.apos.error('notfound');
+          }
+          const replacement = klona(existing);
+          await self.apos.doc.db.removeOne({ _id: from });
+          const oldAposDocId = existing.aposDocId;
+          replacement._id = to;
+          const parts = to.split(':');
+          replacement.aposDocId = parts[0];
+          // Watch out for nonlocalized types, don't set aposLocale for them
+          if (parts.length > 1) {
+            replacement.aposLocale = parts.slice(1).join(':');
+          }
+          const isPage = self.apos.page.isPage(existing);
+          if (isPage) {
+            replacement.path = existing.path.replace(existing.aposDocId, replacement.aposDocId);
+          }
+          try {
+            await self.apos.doc.db.insertOne(replacement);
+          } catch (e) {
+            // First reinsert old doc to prevent content loss on new doc insert failure
+            await self.apos.doc.db.insertOne(existing);
+            if (!self.apos.doc.isUniqueError(e)) {
+              // We cannot fix this error
+              throw e;
+            }
+            const existingReplacement = await self.apos.doc.db.findOne({ _id: replacement._id });
+            if (!existingReplacement) {
+              // We don't know the cause of this error
+              throw e;
+            }
+            if (keep === 'new') {
+              // New content already exists in new locale, delete old locale
+              // and keep new
+              await self.apos.doc.db.removeOne({ _id: existing._id });
+              kept++;
+            } else if (keep === 'old') {
+              // We want to keep the old content, but with the new
+              // identifiers. Once again we need to remove the old doc first
+              // to cut down on conflicts
+              try {
+                await self.apos.doc.db.deleteOne({ _id: existing._id });
+                await self.apos.doc.db.deleteOne({ _id: replacement._id });
+                await self.apos.doc.db.insertOne(replacement);
+              } catch (e) {
+                // Reinsert old doc to prevent content loss on new doc insert failure
+                await self.apos.doc.db.insertOne(existing);
+                throw e;
+              }
+              kept++;
+            } else {
+              throw self.apos.error('conflict');
+            }
+          }
+          if (isPage) {
+            for (const page of pages) {
+              if (page.aposLastTargetId === from) {
+                await self.apos.doc.db.updateOne({
+                  _id: page._id
+                }, {
+                  $set: {
+                    aposLastTargetId: to
+                  }
+                });
+              }
+              if (page.path.includes(oldAposDocId)) {
+                await self.apos.doc.db.updateOne({
+                  _id: page._id
+                }, {
+                  $set: {
+                    path: page.path.replace(oldAposDocId, replacement.aposDocId)
+                  }
+                });
+              }
+            }
+          }
+          if (existing.relatedReverseIds?.length) {
+            const relatedDocs = await self.apos.doc.db.find({
+              aposDocId: { $in: existing.relatedReverseIds }
+            }).toArray();
+            for (const doc of relatedDocs) {
+              replaceId(doc, oldAposDocId, replacement.aposDocId);
+              await self.apos.doc.db.replaceOne({
+                _id: doc._id
+              }, doc);
+            }
+          }
+        }
+        await self.apos.attachment.recomputeAllDocReferences();
+        return {
+          kept
+        };
+        function replaceId(obj, oldId, newId) {
+          if (obj == null) {
+            return;
+          }
+          if ((typeof obj) !== 'object') {
+            return;
+          }
+          for (const key of Object.keys(obj)) {
+            if (obj[key] === oldId) {
+              obj[key] = newId;
+            } else {
+              replaceId(obj[key], oldId, newId);
+            }
+          }
+        }
+      },
       async enableCollection() {
         self.db = await self.apos.db.collection('aposDocs');
       },
@@ -1302,6 +1441,11 @@ module.exports = {
             }
           });
         });
+      },
+
+      async bestAposDocId(criteria) {
+        const existing = await self.apos.doc.db.findOne(criteria, { projection: { aposDocId: 1 } });
+        return existing?.aposDocId || self.apos.util.generateId();
       },
 
       ...require('./lib/legacy-migrations')(self)